Render [[enclosed]] strings as spans
For the first pass at a wordmark, say, I might want to make a certain
word bigger, or italicize a single letter. I previously relied on
`dangerouslySetInnerHTML` for this.

I didn’t like that workflow, so I wrote a quick parser to convert double
brackets into `span`s with unique `className`s. Now all the styling can
happen in devtools, which is a nicer flow for me.

And as a nice bonus – no more `dangerouslySetHTML`, and yet another perf
win :)